+++ src/effects/spectral_arcs.h
@@ -15,10 +15,10 @@ struct SpectralArcsConfig {
   bool enabled = false;
 
   // FFT mapping
-  float baseFreq = 55.0f;  // Lowest visible frequency in Hz (27.5-440.0)
-  float numOctaves = 5.0f; // Octave count (1.0-8.0)
-  float gain = 2.0f;       // FFT magnitude amplifier (0.1-10.0)
-  float curve = 0.7f;      // Contrast exponent on magnitude (0.1-3.0)
+  float baseFreq = 55.0f; // Lowest visible frequency in Hz (27.5-440.0)
+  int numOctaves = 5;     // Octave count (1-8)
+  float gain = 2.0f;      // FFT magnitude amplifier (0.1-10.0)
+  float curve = 0.7f;     // Contrast exponent on magnitude (0.1-3.0)
 
   // Ring layout
   float ringScale = 2.5f; // Ring spread factor (higher = tighter packing)
